% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{assignments}
\alias{assignments}
\title{Extract assignments from a winCurse fit}
\usage{
assignments(x)
}
\arguments{
\item{x}{An object of class \code{winCurse}.}
}
\value{
Data frame with \code{non_null} (0/1) and \code{entropy}.
}
\description{
Extract assignments from a winCurse fit
}
