% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{convergence}
\alias{convergence}
\title{Extract convergence info from a winCurse fit}
\usage{
convergence(x)
}
\arguments{
\item{x}{An object of class \code{winCurse}.}
}
\value{
List with \code{niter} and \code{converged}.
}
\description{
Extract convergence info from a winCurse fit
}
