% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{.ValidatePiTau2}
\alias{.ValidatePiTau2}
\title{Validate pi and tau2 for PostExp}
\usage{
.ValidatePiTau2(pi, tau2)
}
\arguments{
\item{pi}{Proportion null (single number).}

\item{tau2}{Variance component (single number).}
}
\description{
Validate pi and tau2 for PostExp
}
\keyword{internal}
