% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{.ValidateThetaSe}
\alias{.ValidateThetaSe}
\title{Validate theta and se inputs}
\usage{
.ValidateThetaSe(theta, se)
}
\arguments{
\item{theta}{Parameter estimates (numeric vector).}

\item{se}{Parameter standard errors (numeric vector).}
}
\value{
Invisible NULL; stops with an error if invalid.
}
\description{
Validate theta and se inputs
}
\keyword{internal}
