% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{EM.Objective}
\alias{EM.Objective}
\title{EM Objective}
\usage{
EM.Objective(theta, se, pi, tau2, gamma)
}
\arguments{
\item{theta}{Parameter estimates.}

\item{se}{Parameter standard errors.}

\item{pi}{Proportion null parameters.}

\item{tau2}{Variance component.}

\item{gamma}{Responsibilities.}
}
\value{
Numeric EM objective.
}
\description{
EM Objective
}
