% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{EM.Update}
\alias{EM.Update}
\title{EM Update}
\usage{
EM.Update(theta, se, pi, tau2, tau2.upper = 1)
}
\arguments{
\item{theta}{Parameter estimates.}

\item{se}{Parameter standard errors.}

\item{pi}{Proportion null parameters.}

\item{tau2}{Variance component.}

\item{tau2.upper}{Upper bound for tau2 in uniroot (optional).}
}
\value{
List containing:
\itemize{
\item 'pi' updated value of pi.
\item 'tau2' updated value of tau2.
\item 'delta' increment in the EM objective.
}
}
\description{
One iteration of the EM algorithm (E-step and M-step).
}
\keyword{internal}
