% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{estimates}
\alias{estimates}
\title{Extract estimates from a winCurse fit}
\usage{
estimates(x)
}
\arguments{
\item{x}{An object of class \code{winCurse}.}
}
\value{
List with \code{pi} and \code{tau2}.
}
\description{
Extract estimates from a winCurse fit
}
