% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{expectations}
\alias{expectations}
\title{Extract posterior expectations from a winCurse fit}
\usage{
expectations(x)
}
\arguments{
\item{x}{An object of class \code{winCurse}.}
}
\value{
Numeric vector of posterior expected effect sizes.
}
\description{
Extract posterior expectations from a winCurse fit
}
