% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{show,winCurse-method}
\alias{show,winCurse-method}
\title{Show Method for Winner's Curse Object}
\usage{
\S4method{show}{winCurse}(object)
}
\arguments{
\item{object}{An object of class \code{winCurse}.}
}
\description{
Show Method for Winner's Curse Object
}
