% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{print.winCurse}
\alias{print.winCurse}
\title{Print Method for Winner's Curse Object.}
\usage{
\method{print}{winCurse}(x, ...)
}
\arguments{
\item{x}{An object of class \code{winCurse}.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{winCurse}.
}
