% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{responsibilities}
\alias{responsibilities}
\title{Extract responsibilities from a winCurse fit}
\usage{
responsibilities(x)
}
\arguments{
\item{x}{An object of class \code{winCurse}.}
}
\value{
Data frame with columns \code{null} and \code{non_null}.
}
\description{
Extract responsibilities from a winCurse fit
}
