% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{Responsibility}
\alias{Responsibility}
\title{Responsibility}
\usage{
Responsibility(theta, se, pi, tau2)
}
\arguments{
\item{theta}{Parameter estimates.}

\item{se}{Parameter standard errors.}

\item{pi}{Proportion null parameters.}

\item{tau2}{Variance component.}
}
\value{
Vector of responsibilities.
}
\description{
Responsibility
}
