% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-data.R
\docType{data}
\name{wc_data}
\alias{wc_data}
\title{Sample Winners Curse Data}
\format{
A data.frame containing four fields:
\describe{
\item{non_null}{Generative component, 1 if non-null, 0 otherwise.}
\item{theta}{Parameter estimate.}
\item{se}{Parameter standard error.}
}
}
\usage{
data(wc_data)
}
\description{
Example data for winner's curse correction.
}
\keyword{datasets}
