% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{winCurse-class}
\alias{winCurse-class}
\title{Winner's Curse Object}
\description{
Defines the object class returned by \code{\link{fit.WinCurse}}.
}
\section{Slots}{

\describe{
\item{\code{Assignments}}{Maximum a posteriori component assignment
and assignment entropy.}

\item{\code{Estimates}}{Estimated model parameters.}

\item{\code{Expectations}}{Posterior expected effect sizes.}

\item{\code{Responsibilities}}{Probabilities the parameter came from the null and
non-null components.}
}}

