% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{EM.Update}
\alias{EM.Update}
\title{EM Update}
\usage{
EM.Update(theta, se, pi, tau2)
}
\arguments{
\item{theta}{Parameter estimates.}

\item{se}{Parameter standard errors.}

\item{pi}{Proportion null parameters.}

\item{tau2}{Variance component.}
}
\value{
List containing:
\itemize{
\item 'pi' updated value of pi.
\item 'tau2' updated value of tau2.
\item 'delta' increment in the EM objective.
}
}
\description{
EM Update
}
