% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{fit.WinCurse}
\alias{fit.WinCurse}
\title{Fit Winner's Curse Model}
\usage{
fit.WinCurse(
  theta,
  se,
  pi = NULL,
  tau2 = NULL,
  eps = 1e-06,
  maxit = 100,
  report = FALSE
)
}
\arguments{
\item{theta}{Parameter estimates.}

\item{se}{Parameter standard errors.}

\item{pi}{Initial value of pi, the proportion of null parameters.}

\item{tau2}{Initial value of tau, the variance component.}

\item{eps}{Tolerance for Newton-Raphson iterations.}

\item{maxit}{Maximum number of NR iterations.}

\item{report}{Report fitting progress?}
}
\value{
Object of class 'winCurse' containing:
\itemize{
\item \verb{@Assignments}: Maximum a posteriori component assignment
and assignment entropy.
\item \verb{@Estimates}: Estimated model parameters.
\item \verb{@Expectations}: Posterior expected effect sizes.
\item \verb{@Responsibilities}: Probabilities the parameter came from the null and
non-null components.
}
}
\description{
Fit Winner's Curse Model
}
\examples{
data(wc_data)
fit <- fit.WinCurse(
  theta = wc_data$theta,
  se = wc_data$se
)
}
