% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{PostExp}
\alias{PostExp}
\title{Posterior Expectation}
\usage{
PostExp(theta, se, pi, tau2)
}
\arguments{
\item{theta}{Parameter estimates.}

\item{se}{Parameter standard errors.}

\item{pi}{Initial value of pi, the proportion of null parameters.}

\item{tau2}{Initial value of tau, the variance component.}
}
\value{
Numeric vector of the posterior expected effect sizes.
}
\description{
Calculates the posterior expectation given available \code{pi} and \code{tau2}.
}
\examples{
data(wc_data)
post_exp <- PostExp(
  theta = wc_data$theta,
  se = wc_data$se,
  pi = 0.75,
  tau2 = 0.05
)
}
